переделал store на slice'ах.
