--- src/store/actions.ts
@@ -1,3 +1,0 @@
-import { createAction } from '@reduxjs/toolkit';
-
-export const changeCity = createAction<string>('cards/changeCity');
