+++ src/store/hooks.ts
@@ -1,7 +1,6 @@
 import { TypedUseSelectorHook, useSelector } from 'react-redux';
 import { useDispatch } from 'react-redux';
-
-import { AppDispatch, AppState } from '@store/types';
+import { AppDispatch, AppState } from './store';
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 export const useAppSelector: TypedUseSelectorHook<AppState> = useSelector;
