+++ src/store/store.ts
@@ -1,4 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { reducer } from './reducer';
+import reducer from './cardsSlice';
 
 export const store = configureStore({ reducer });
+
+export type AppState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
