--- src/store/types.ts
@@ -1,5 +1,0 @@
-import { store } from './store';
-
-export type AppState = ReturnType<typeof store.getState>;
-
-export type AppDispatch = typeof store.dispatch;
